file_project/requirements.txt
asgiref==3.4.1
Django==3.2.7
Pillow==8.3.2
pip==21.2.4
pytz==2021.1
setuptools==58.0.4
sqlparse==0.4.1

